Update some loop vars to be size_t to remove warnings, caused by switching Map type counts to been size_t: 0f2dcacaa0e6bc12359ee34f81a1b5c790cd7910

+++ tools/LevelBuilder/LvlBuilder.cpp
@@ -262,7 +262,7 @@ int32_t LvlBuilder::FindFloatPlane(const Planef& plane)
 bool LvlBuilder::processMapEntity(LvlEntity& ent, mapfile::XMapEntity* mapEnt)
 {
 	mapfile::XMapPrimitive* prim;
-	int32_t i;
+	size_t i;
 
 	// update stats.
 	stats_.numEntities++;
